# account-service/target/generated-sources/openapi/src/main/java/com/nttdata/accountservice/generated/api/AccountsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.5.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.nttdata.accountservice.generated.api;

import com.nttdata.accountservice.generated.model.AccountResponse;
import com.nttdata.accountservice.generated.model.BalanceOperationRequest;
import com.nttdata.accountservice.generated.model.CreateAccountRequest;
import com.nttdata.accountservice.generated.model.UpdateAccountRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.server.ServerWebExchange;
import reactor.core.publisher.Flux;
import reactor.core.publisher.Mono;
import org.springframework.http.codec.multipart.Part;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-03-23T20:37:50.755963500-05:00[America/Lima]", comments = "Generator version: 7.5.0")
@Validated
@Tag(name = "Accounts", description = "Account operations")
public interface AccountsApi {

    default AccountsApiDelegate getDelegate() {
        return new AccountsApiDelegate() {};
    }

    /**
     * POST /api/accounts : Create account
     *
     * @param createAccountRequest  (required)
     * @return Account created (status code 201)
     */
    @Operation(
        operationId = "createAccount",
        summary = "Create account",
        tags = { "Accounts" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Account created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AccountResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/accounts",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default Mono<ResponseEntity<AccountResponse>> createAccount(
        @Parameter(name = "CreateAccountRequest", description = "", required = true) @Valid @RequestBody Mono<CreateAccountRequest> createAccountRequest,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return getDelegate().createAccount(createAccountRequest, exchange);
    }


    /**
     * DELETE /api/accounts/{accountId} : Delete account
     *
     * @param accountId  (required)
     * @return Account deleted (status code 204)
     */
    @Operation(
        operationId = "deleteAccount",
        summary = "Delete account",
        tags = { "Accounts" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Account deleted")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/api/accounts/{accountId}"
    )
    
    default Mono<ResponseEntity<Void>> deleteAccount(
        @Parameter(name = "accountId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("accountId") String accountId,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return getDelegate().deleteAccount(accountId, exchange);
    }


    /**
     * GET /api/accounts/{accountId} : Get account by id
     *
     * @param accountId  (required)
     * @return Account found (status code 200)
     */
    @Operation(
        operationId = "getAccountById",
        summary = "Get account by id",
        tags = { "Accounts" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Account found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AccountResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/accounts/{accountId}",
        produces = { "application/json" }
    )
    
    default Mono<ResponseEntity<AccountResponse>> getAccountById(
        @Parameter(name = "accountId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("accountId") String accountId,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return getDelegate().getAccountById(accountId, exchange);
    }


    /**
     * GET /api/accounts : List accounts
     *
     * @param customerId  (optional)
     * @return Account list (status code 200)
     */
    @Operation(
        operationId = "getAccounts",
        summary = "List accounts",
        tags = { "Accounts" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Account list", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = AccountResponse.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/accounts",
        produces = { "application/json" }
    )
    
    default Mono<ResponseEntity<Flux<AccountResponse>>> getAccounts(
        @Parameter(name = "customerId", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "customerId", required = false) String customerId,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return getDelegate().getAccounts(customerId, exchange);
    }


    /**
     * POST /api/accounts/{accountId}/operations : Apply balance operation
     *
     * @param accountId  (required)
     * @param balanceOperationRequest  (required)
     * @return Operation applied (status code 200)
     */
    @Operation(
        operationId = "operateBalance",
        summary = "Apply balance operation",
        tags = { "Accounts" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Operation applied", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AccountResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/accounts/{accountId}/operations",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default Mono<ResponseEntity<AccountResponse>> operateBalance(
        @Parameter(name = "accountId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("accountId") String accountId,
        @Parameter(name = "BalanceOperationRequest", description = "", required = true) @Valid @RequestBody Mono<BalanceOperationRequest> balanceOperationRequest,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return getDelegate().operateBalance(accountId, balanceOperationRequest, exchange);
    }


    /**
     * PUT /api/accounts/{accountId} : Update account
     *
     * @param accountId  (required)
     * @param updateAccountRequest  (required)
     * @return Account updated (status code 200)
     */
    @Operation(
        operationId = "updateAccount",
        summary = "Update account",
        tags = { "Accounts" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Account updated", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AccountResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/api/accounts/{accountId}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default Mono<ResponseEntity<AccountResponse>> updateAccount(
        @Parameter(name = "accountId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("accountId") String accountId,
        @Parameter(name = "UpdateAccountRequest", description = "", required = true) @Valid @RequestBody Mono<UpdateAccountRequest> updateAccountRequest,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return getDelegate().updateAccount(accountId, updateAccountRequest, exchange);
    }

}